+++ src/main/java/com/keepu/webAPI/dto/response/InvitationCodeResponse.java
@@ -3,11 +3,11 @@ package com.keepu.webAPI.dto.response;
 import java.time.LocalDateTime;
 
 public record InvitationCodeResponse(
-        Integer id,
+        Long id,
         String code,
         boolean isUsed,
         LocalDateTime expiresAt,
-        Integer userId,
+        Long userId,
         String childName,
         String childLastName
 ) {}
