+++ src/main/java/com/keepu/webAPI/repository/ParentRepository.java
@@ -3,6 +3,6 @@ package com.keepu.webAPI.repository;
 import com.keepu.webAPI.model.Parent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParentRepository extends JpaRepository<Parent, Integer> {
-    Parent findByUserId(Integer userId);
+public interface ParentRepository extends JpaRepository<Parent, Long> {
+    Parent findByUserId(Long userId);
 }
